pythonDataSync/requirements.txt
pandas
openpyxl
sqlalchemy
mysql-connector-python
python-dotenv
pydantic
fastapi
uvicorn
python-multipart


